requirements.txt
Package           Version
----------------- -------
annotated-types   0.7.0
anyio             4.9.0
click             8.2.1
colorama          0.4.6
fastapi           0.116.1
h11               0.16.0
idna              3.10
iniconfig         2.1.0
packaging         25.0
pip               25.1.1
pluggy            1.6.0
prometheus_client 0.22.1
pydantic          2.11.7
pydantic_core     2.33.2
Pygments          2.19.2
pytest            8.4.1
PyYAML            6.0.2
sniffio           1.3.1
starlette         0.47.2
typing_extensions 4.14.1
typing-inspection 0.4.1
uvicorn           0.35.0
